requirements.txt
pydantic==2.11.5
databricks-agents==0.22.1
mlflow>=2.18.0
databricks-sdk[openai]==0.56.0
unitycatalog-langchain[databricks]==0.2.0
unitycatalog-ai[databricks]==0.3.1
pyyaml==6.0.2
tabulate==0.9.0
pandas==2.2.3
databricks-connect==16.1.5
Flask==2.0.1
openai==1.83.0